pyproject.toml
[project]
name = "drawing-llm"
version = "0.1.0"
description = "The kaggle project of drawing llm"
authors = [
    {name = "Linghua Zhang",email = "lz101@rice.edu"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "openai (>=1.66.3,<2.0.0)",
    "cairosvg (>=2.7.1,<3.0.0)",
    "google-re2 (>=1.1.20240702,<2.0.0)",
    "kagglehub (>=0.3.10,<0.4.0)",
    "lxml (>=5.3.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
